批量结算

+++ src/main/java/com/yt/app/api/v1/vo/IncomeVO.java
@@ -38,7 +38,7 @@ public class IncomeVO extends BaseVO {
 	Long agentid;
 	String agentordernum;
 	Double agentincome;
-	Double qrcodeid;
+	Long qrcodeid;
 	String qrcodename;
 	String qrcodecode;
 	String qrcodeaislecode;
